Cast atoi result to byte.

+++ g_svcmds.c
@@ -95,7 +95,7 @@ qboolean StringToFilter(char* s, ipfilter_t* f)
 			num[j++] = *s++;
 		}
 		num[j] = 0;
-		b[i] = atoi(num);
+		b[i] = (byte)atoi(num);
 
 		if (!*s)
 		{
